:truck: magic renaming
